REF: move platform blacklist into its own module.
